feat(status): sync with replicaset's event
